Moved heartbeat failure counter
